Module Morph__.Router

type route = Morph.Server.handler Routes.route

val make : ?⁠get:route list -> ?⁠post:route list -> ?⁠put:route list -> ?⁠del:route list -> ?⁠print:bool -> ?⁠not_found_handler:Morph.Server.handler -> unit -> Morph.Server.handler

make takes lists of routes for each method and returns a Morph.Server.handler

print will defaults to true which will print the routing table

Sample usage:

let get_routes =
  Routes.[
    s("greet") /? nil @--> greet_handler,
    s("set_greeting") / str /? nil @--> set_greeting_handler,
    s("secure") /? nil @--> secured_handler,
  ];

let handler = Morph.Router.make(~get=get_routes, ());

let server =
  Morph.Server.make(~port=3333, ~address=Unix.inet_addr_loopback, ());

server.start(handler) |> Lwt_main.run

Output:

Routing table

GET
/greet
/secure
/set_greeting/:string