Module Morph__Method

type standard = [

| `GET
RFC72314.3.1. Safe, Cacheable.

| `HEAD
RFC72314.3.2. Safe, Cacheable.

| `POST
RFC72314.3.3. Cacheable.

| `PUT
RFC72314.3.4. Idempotent.

| `DELETE
RFC72314.3.5. Idempotent.

| `CONNECT
RFC72314.3.6.

| `OPTIONS
RFC72314.3.7. Safe.

| `TRACE
RFC72314.3.8. Safe.

];

type t = [

| standard

| `Other(string)
Methods defined outside of RFC7231, or custom methods.

];

let is_safe: standard => bool;

Request methods are considered "safe" if their defined semantics are essentially read-only; i.e., the client does not request, and does not expect, any state change on the origin server as a result of applying a safe method to a target resource. Likewise, reasonable use of a safe method is not expected to cause any harm, loss of property, or unusual burden on the origin server. See RFC72314.2.1 for more details.

let is_cacheable: standard => bool;

Request methods can be defined as "cacheable" to indicate that responses to them are allowed to be stored for future reuse. See RFC7234 for more details.

let is_idempotent: standard => bool;

A request method is considered "idempotent" if the intended effect on the server of multiple identical requests with that method is the same as the effect for a single such request. Of the request methods defined by this specification, PUT, DELETE, and safe request methods are idempotent. See RFC72314.2.2 for more details.

let to_string: t => string;

let of_string: string => t;

let pp_hum: Stdlib.Format.formatter => t => unit;