Morph client

A HTTP client built ontop of Httpaf that leverages Opium_core to build reusable middlewares.

The entry point of this library is the module: Morph_client.

Core concepts

Morph_client is built on top of Opium_core that provides us with Filters and Services. The Morph_client is just a Opium_core.Serivce.t so it takes a Morph.Request.t and returns a Morph.Response.t wrapped in a Lwt.t.

To read more about Filters and Services I recommend going to the Opium_core documentation.

For examples go to the reason-native-web homepage or look in the examples folder of this project.