Package morph_graphql_server 0.4.0

The entry point of this library is the module: Morph_graphql_server.

Goals

This package aims to be a simple way of building graphql servers. You should just have to build your schema and then have it running.

Package info

authors

Ulrik Strid

changes-files

CHANGES.md

depends

dune

graphql

graphql-lwt

logs

lwt

morph

ocaml

reason

homepage

https://reason-native-web.github.io

issues

https://github.com/reason-native-web/morph/issues

maintainers

Ulrik Strid

online-doc

https://reason-native-web.github.io/morph

readme-files

README.md

repo

git+https://github.com/reason-native-web/morph.git

version

0.4.0